refactor signals
