+++ puzzles/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class PuzzlesConfig(AppConfig):
     name = 'puzzles'
+
+    def ready(self):
+        import puzzles.signals.handlers
