+++ smallboard/settings.py
@@ -56,7 +56,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'puzzles',
+    'puzzles.apps.PuzzlesConfig',
     'accounts',
     'hunts',
     'answers',
